+++ relm4-macros/src/widgets/gen/util/mod.rs
@@ -1,3 +1,4 @@
+mod has_struct_field;
 mod if_branch;
 mod property_name;
 mod widget_func;