+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
++ macros: Don't generate dead code in the widgets struct
 + macros: Improve error reporting on invalid trait implementations
 + macros: Add chain attribute for properties
 + core: Add `iter_mut` to `FactoryVecDeque`